feat: CMakeLists to build root project and tests. clang-format intergration. Added README file. Added ./build and ./src folders
