+++ .clang-format
@@ -1,0 +1,1 @@
+IndentWidth: 4
