+++ CMakeLists.txt
@@ -1,0 +1,38 @@
+cmake_minimum_required(VERSION 3.14)
+project(fluid_dynamics)
+
+set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
+
+# NOTE: To be honest don't know if it's a good solution (:
+file(GLOB all_SRCS
+        "src/*.h"
+        "src/*.hpp"
+        "src/*.cpp"
+        "src/*.c"
+)
+
+add_executable(fluid_dynamics src/main.cpp)
+
+include(FetchContent)
+FetchContent_Declare(
+	googletest
+	URL https://github.com/google/googletest/archive/03597a01ee50ed33e9dfd640b249b4be3799d395.zip
+)
+# For Windows: Prevent overriding the parent project's compiler/linker settings
+set(gtest_force_shared_crt ON CACHE BOOL "" FORCE)
+FetchContent_MakeAvailable(googletest)
+
+enable_testing()
+
+add_executable(
+	tests
+	src/tests.cpp
+)
+target_link_libraries(
+	tests
+	GTest::gtest_main
+)
+
+include(GoogleTest)
+gtest_discover_tests(tests)
