--- main.cpp
@@ -1,7 +1,0 @@
-#include"types.h"
-
-int main(){
-
-    
-    return 0;
-}
